@@ -1,3 +1,3 @@
 module test
 
-go 1.21
+go 1.23.0
