@@ -12,7 +12,7 @@ func TestGenPrimes(t *testing.T) {
 	primes20 := []int{2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71}
 
 	var got []int
-	for p := range genPrimes {
+	for p := range genPrimes() {
 		got = append(got, p)
 		if len(got) >= 20 {
 			break
@@ -20,7 +20,7 @@ func TestGenPrimes(t *testing.T) {
 	}
 
 	var gotOpt []int
-	for p := range genPrimesOpt {
+	for p := range genPrimesOpt() {
 		gotOpt = append(gotOpt, p)
 		if len(gotOpt) >= 20 {
 			break
@@ -37,9 +37,9 @@ func TestGenPrimes(t *testing.T) {
 
 func BenchmarkGenPrimes(b *testing.B) {
 	result := 0
-	for i := 0; i < b.N; i++ {
+	for range b.N {
 		n := 0
-		for p := range genPrimes {
+		for p := range genPrimes() {
 			result += p
 			n++
 			if n >= 9000 {
@@ -52,9 +52,9 @@ func BenchmarkGenPrimes(b *testing.B) {
 
 func BenchmarkGenPrimesOpt(b *testing.B) {
 	result := 0
-	for i := 0; i < b.N; i++ {
+	for range b.N {
 		n := 0
-		for p := range genPrimesOpt {
+		for p := range genPrimesOpt() {
 			result += p
 			n++
 			if n >= 9000 {
